requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
azure-search-documents==11.5.2
azure-core==1.32.0
click==8.1.7
colorama==0.4.6
fastapi==0.115.5
greenlet==3.1.1
h11==0.14.0
idna==3.10
openai==1.65.5
pandas==2.2.3
pydantic==2.9.2
pydantic_core==2.23.4
PyMySQL==1.1.1
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
mysql-connector-python==9.1.0
uvicorn==0.32.0
gunicorn[uvicorn]
bcrypt==4.3.0
